updated add to cart functionality

+++ src/main/java/com/progetto/BookHavenBackend/services/BookService.java
@@ -1,13 +1,17 @@
 package com.progetto.BookHavenBackend.services;
 
 import com.progetto.BookHavenBackend.entities.Book;
+import com.progetto.BookHavenBackend.entities.Inventory;
 import com.progetto.BookHavenBackend.repositories.BookRepository;
+import com.progetto.BookHavenBackend.repositories.InventoryRepository;
 import com.progetto.BookHavenBackend.support.exceptions.BookNotFoundException;
+import com.progetto.BookHavenBackend.support.exceptions.CustomException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,11 +20,13 @@ import java.util.Optional;
 public class BookService {
     @Autowired
     BookRepository bookRepository;
+    @Autowired
+    private InventoryRepository inventoryRepository;
+
     public List<Book> getAllBooks() {
         return bookRepository.findAll();
     }
 
-    @Transactional(readOnly = true)
     public Book getBookById(long bookId) throws BookNotFoundException {
         Optional<Book> bookOptional = Optional.ofNullable(bookRepository.findBookById(bookId));
         if( bookOptional.isPresent()){
@@ -42,8 +48,7 @@ public class BookService {
 
     public List<Book> getBestSellingBooks() {
         List<Book> bestSellers = bookRepository.sortByNumberOfBuyers();
-        // If there are more than 10 best sellers, return the first 10.
-        return bestSellers.size() > 10 ? bestSellers.subList(0, 10) : bestSellers;
+        return bestSellers.size() > 10 ? bestSellers.subList(0, 10) : bestSellers; // If there are more than 10 best sellers, return the first 10.
     }
 
     public Book addNewBook(Book book) {
@@ -63,23 +68,28 @@ public class BookService {
 
     public void deleteBookById(long id) throws BookNotFoundException {
         Optional<Book> bookOptional = Optional.ofNullable(bookRepository.findBookById(id));
-        if( bookOptional.isPresent()){
-            bookRepository.deleteById(id);
+        if( bookOptional.isPresent() ){
+            if (inventoryRepository.findById(id).get().getQuantity() == 0) {
+                bookRepository.deleteById(id);
+            } else {
+                throw new CustomException("Book can not be deleted because it is not empty in inventory.");
+            }
         }else{
             throw new BookNotFoundException();
         }
     }
 
-    public List<Book> getWishlist() {
-        List<Book> wishlist = bookRepository.findBooksInWishlist();
-        return wishlist;
-    }
-
-    public Book findBookById(Long bookId) throws BookNotFoundException {
-        Optional<Book> optionalBook = bookRepository.findById(bookId);
-        if (optionalBook.isEmpty()){
-            throw new BookNotFoundException("Product id is not valid" + bookId);
+    public void updateBookQuantityInInventory(Long id, Long quantity) throws BookNotFoundException {
+        Optional<Book> bookOptional = Optional.ofNullable(bookRepository.findBookById(id));
+        if (bookOptional.isPresent()){
+            Inventory bookInventory = new Inventory();
+            bookInventory.setId(id);
+            bookInventory.setQuantity(quantity);
+            bookInventory.setDateTimeBookIsLastUpdated(LocalDateTime.now());
+            inventoryRepository.save(bookInventory);
+        }else{
+            throw new BookNotFoundException();
         }
-        return optionalBook.get();
+
     }
 }
